Warning with GCC 4.5: ../src/btree/row_put.c: In function '__wt_repl_alloc': ../src/btree/row_put.c:171:24: warning: comparison of promoted ~unsigned with constant ../src/btree/row_put.c:175:30: warning: comparison of promoted ~unsigned with constant

+++ examples/c/ex_extending.c
@@ -39,7 +39,7 @@ static WT_COLLATOR nocasecoll = { __compare_nocase };
  */
 typedef struct {
 	WT_COLLATOR iface;
-	int maxlen;
+	size_t maxlen;
 } PREFIX_COLLATOR;
 
 static int
